C#; Software Testing class final examination
Test #1: Fill the CoffeeMaker using its Fill method. Make a Grande(3) using MakeCoffee. Assert that both water and coffee volumes are now 7 using the CoffeeVolume and WaterVolume properties.

Test #2: Fill the coffee to 8 using the CoffeeVolume property. Fill the water to 6 using the WaterVolume property. Make a Tall(2) using MakeCoffee. Assert that water volume is now 4 using the WaterVolume property and assert that coffee is now 6 using CoffeeVolume.

Test #3: Set the coffee and water volumes to 5 using CoffeeVolume and WaterVolume. Clean() the CoffeeMaker. Assert that the coffee and water volumes are now 0 using the IsEmpty property.

Test #4: Set the coffee volume to 4 using CoffeeVolume. Set the water volume to 2 using WaterVolume. Make a Grande(3) using MakeCoffee. Check that the InvalidOperationException is thrown.

Test #5 (Moq only!): Clean() the CoffeeMaker(). Verify that the water container's Clean() was called once. Verify that the coffee containers Clean() was called twice.